ai_compliance_agent/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.32
langchain-core>=0.1.20
langchain-classic>=0.0.27
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.3
langchain-ollama>=0.1.3
requests>=2.31.0
python-dotenv>=1.0.0
pypdf>=3.10.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
gradio>=3.50.0
huggingface-hub>=0.20.0
torch>=2.0.0
